Split Glitter in sub-files
Was just getting too big of a file!
