+++ interfaces/Glitter/templates/main.tmpl
@@ -127,7 +127,7 @@
             <!--#include raw $webdir + "/static/javascripts/knockout-latest.js"#-->
             <!--#include raw $webdir + "/static/javascripts/knockout-extensions.js"#-->
             <!--#include raw $webdir + "/static/bootstrap/js/bootstrap.min.js"#-->
-            <!--#include raw $webdir + "/static/javascripts/glitter.js"#-->
+            <!--#include $webdir + "/static/javascripts/glitter.js"#-->
         </script>
         <!--#if active_lang != 'en'#-->
         <script type="text/javascript" src="./static/javascripts/momentjs_locale/${active_lang}.js"></script>
